requirements.txt
fastapi==0.115.12
fuzzywuzzy==0.18.0
flask==3.1.1
python_Levenshtein==0.27.1
gradio==5.31.0
numpy==2.2.6
openai==1.82.0
pydantic==2.11.5
pydantic_settings==2.9.1
python-dotenv==1.1.0
python_dateutil==2.9.0.post0
Requests==2.32.3
sentence_transformers==4.1.0
tenacity==9.0.0
tiktoken==0.9.0
openai_agents==0.0.16
